requirements.txt
Patrowl4py==1.0.2
python-dateutil==2.7.2
requests==2.21.0
